fix visualizer
