+++ Engine/Physics/ColliderNode/ColliderNode.cs
@@ -8,8 +8,6 @@ namespace Altseed2
     [Serializable]
     public abstract class ColliderNode : TransformNode
     {
-        internal readonly static Color AreaColor = new Color(255, 100, 100, 100);
-
         public override Matrix44F AbsoluteTransform => _absoluteTransform;
         private Matrix44F _absoluteTransform;
 
@@ -22,24 +20,6 @@ namespace Altseed2
         /// <see cref="ColliderNode"/>の新しいインスタンスを生成する
         /// </summary>
         private protected ColliderNode() { }
-
-        /// <summary>
-        /// 指定した<see cref="ColliderNode"/>の当たり判定領域を表示するノードを生成する
-        /// </summary>
-        /// <param name="colliderNode">使用するコライダノード</param>
-        /// <exception cref="ArgumentNullException"><paramref name="colliderNode"/>がnull</exception>
-        /// <returns><paramref name="colliderNode"/>の当たり当たり領域を表示するノード</returns>
-        public static ColliderVisualizeNode CreateVisualizeNode(ColliderNode colliderNode)
-        {
-            if (colliderNode == null) throw new ArgumentNullException(nameof(colliderNode), "引数がnullです");
-            return colliderNode switch
-            {
-                CircleColliderNode c => new CircleColliderNode.CircleColliderVisualizeNode(c),
-                PolygonColliderNode p => new PolygonColliderNode.PolygonColliderVisualizeNode(p),
-                RectangleColliderNode r => new RectangleColliderNode.RectangleColliderVisualizeNode(r),
-                _ => throw new ArgumentException($"サポートされていない型です\n型：{colliderNode.GetType()}", nameof(colliderNode))
-            };
-        }
 
         private static CollisionManagerNode SearchManagerFromChildren(Node node)
         {
